feat: add tutorial data type and implement filter components
- Created a new type  in  for fetching tutorial data with related chapters and categories.
- Implemented  component for mobile filter UI, allowing users to select filter tabs.
- Developed  component for displaying filter options with toggle functionality.
- Added utility function  in  to manage URL search parameters.
- Established a Zustand store  to manage filter tabs and active tab state.
- Introduced  component in the UI library for consistent alert dialog presentation across the application.

+++ apps/admin/src/actions/tutorials/chapters.ts
@@ -88,8 +88,64 @@ export const createChapter = async (
   }
 };
 
+export const editTutorialChapter = async (
+  values: z.infer<typeof tutorialChapterSchema>,
+  chapterId: string,
+) => {
+  try {
+    if (!chapterId) {
+      return {
+        success: false,
+        message: 'Chapter id is required for updating.',
+      };
+    }
+
+    const existingChapter = await prisma.chapter.findUnique({
+      where: { id: chapterId },
+    });
+
+    if (!existingChapter) {
+      return {
+        success: false,
+        message: 'chapter not found.',
+      };
+    }
+
+    await prisma.chapter.update({
+      where: { id: chapterId },
+      data: {
+        ...values,
+      },
+    });
+
+    return {
+      success: true,
+      message: 'Chapter Updated Successfully!',
+    };
+  } catch (error) {
+    console.error(error);
+    return {
+      success: false,
+      message: 'Failed to update chapter!',
+    };
+  }
+};
+
 export const deleteTutorialChapter = async (id: string) => {
   try {
+    const chapter = await prisma.chapter.findUnique({
+      where: {
+        id,
+      },
+      include: {
+        _count: true,
+      },
+    });
+
+    if (chapter?._count.topics && chapter._count.topics > 0) {
+      return { success: false, message: 'Chapter contains topics delete them first' };
+    }
+
     await prisma.chapter.delete({
       where: {
         id,
